tasks/sendCommand into main (#9)
* Edit layouts

* Send custom command depending on device type

* Make cards update every 500ms and only update changing ones

* Change color

+++ app/src/main/java/com/persello/domotics/data/device/DeviceData.kt
@@ -4,7 +4,7 @@ data class DeviceData(
     val id: String,
     val type: String,
     val availableCommands: List<String>,
-    val opening: Int,
+    val opening: Double,
     val openingMode: Int,
     val power: Int,
 )
